+++ tests/test_go_server.py
@@ -27,36 +27,6 @@ def test_parser_go_server_ping_no_metrics(tmp_path, capsys):
     assert all(False for _ in tmp_path.iterdir())
 
 
-def test_parser_go_server_ping_file(tmp_path, capsys):
-    """Test that no files are generated if ping definitions
-    are provided."""
-    translate.translate(
-        [
-            ROOT / "data" / "server_metrics_with_event.yaml",
-            ROOT / "data" / "server_pings.yaml",
-        ],
-        "go_server",
-        tmp_path,
-    )
-    assert all(False for _ in tmp_path.iterdir())
-
-
-def test_parser_go_server_metrics_no_ping(tmp_path, capsys):
-    """Test that no files are generated if only metric definitions
-    are provided without any events metrics."""
-    translate.translate(
-        ROOT / "data" / "server_metrics_no_events_no_pings.yaml",
-        "go_server",
-        tmp_path,
-    )
-
-    captured = capsys.readouterr()
-    assert all(False for _ in tmp_path.iterdir())
-    assert (
-        "No event metrics found...at least one event metric is required" in captured.out
-    )
-
-
 def test_parser_go_server_metrics_unsupported_type(tmp_path, capsys):
     """Test that no files are generated with unsupported metric types."""
     translate.translate(
@@ -81,10 +51,10 @@ def test_parser_go_server_metrics_unsupported_type(tmp_path, capsys):
         assert t in captured.out
 
 
-def test_parser_go_server(tmp_path):
-    """Test that parser works"""
+def test_parser_go_server_events_only(tmp_path):
+    """Test that parser works for definitions that only use events ping"""
     translate.translate(
-        ROOT / "data" / "go_server_metrics.yaml",
+        ROOT / "data" / "go_server_events_only_metrics.yaml",
         "go_server",
         tmp_path,
     )
@@ -94,7 +64,61 @@ def test_parser_go_server(tmp_path):
     # Make sure generated file matches expected
     with (tmp_path / "server_events.go").open("r", encoding="utf-8") as fd:
         content = fd.read()
-        with (ROOT / "data" / "server_events_compare.go").open(
+        with (ROOT / "data" / "server_events_only_compare.go").open(
+            "r", encoding="utf-8"
+        ) as cd:
+            compare_raw = cd.read()
+
+    glean_version = f"glean_parser v{glean_parser.__version__}"
+    # use replace instead of format since Go uses { }
+    compare = compare_raw.replace("{current_version}", glean_version)
+    assert content == compare
+
+
+def test_parser_go_server_events_and_custom_ping(tmp_path):
+    """Test that parser works for definitions that use events ping and custom pings"""
+    translate.translate(
+        [
+            ROOT / "data" / "go_server_events_and_custom_ping_metrics.yaml",
+            ROOT / "data" / "go_server_events_and_custom_ping_pings.yaml"
+        ],
+        "go_server",
+        tmp_path,
+    )
+
+    assert set(x.name for x in tmp_path.iterdir()) == set(["server_events.go"])
+
+    # Make sure generated file matches expected
+    with (tmp_path / "server_events.go").open("r", encoding="utf-8") as fd:
+        content = fd.read()
+        with (ROOT / "data" / "server_events_and_custom_ping_compare.go").open(
+            "r", encoding="utf-8"
+        ) as cd:
+            compare_raw = cd.read()
+
+    glean_version = f"glean_parser v{glean_parser.__version__}"
+    # use replace instead of format since Go uses { }
+    compare = compare_raw.replace("{current_version}", glean_version)
+    assert content == compare
+
+
+def test_parser_go_server_custon_ping_only(tmp_path):
+    """Test that parser works for definitions that only use custom pings"""
+    translate.translate(
+        [
+            ROOT / "data" / "go_server_custom_ping_only_metrics.yaml",
+            ROOT / "data" / "go_server_custom_ping_only_pings.yaml"
+        ],
+        "go_server",
+        tmp_path,
+    )
+
+    assert set(x.name for x in tmp_path.iterdir()) == set(["server_events.go"])
+
+    # Make sure generated file matches expected
+    with (tmp_path / "server_events.go").open("r", encoding="utf-8") as fd:
+        content = fd.read()
+        with (ROOT / "data" / "server_custom_ping_only_compare.go").open(
             "r", encoding="utf-8"
         ) as cd:
             compare_raw = cd.read()
@@ -128,27 +152,31 @@ def run_logger(code_dir, code):
 
 
 @pytest.mark.go_dependency
-def test_run_logging(tmp_path):
+def test_run_logging_events_ping(tmp_path):
     glean_module_path = tmp_path / "glean"
 
     translate.translate(
         [
-            ROOT / "data" / "server_metrics_with_event.yaml",
+            ROOT / "data" / "go_server_events_only_metrics.yaml",
         ],
         "go_server",
         glean_module_path,
     )
 
     code = """
-    logger.RecordEventBackendObjectUpdate(
+    logger.RecordEventBackendTestEvent(
         glean.RequestInfo{
             UserAgent: "glean-test/1.0",
             IpAddress: "127.0.0.1",
         },
-        glean.EventBackendObjectUpdate{
-            IdentifiersFxaAccountId: "accountId",
-            ObjectType:              "type",
-            ObjectState:             "state",
+        glean.EventBackendTestEvent{
+            MetricName:            "string value",
+            MetricRequestBool:     true,
+            MetricRequestCount:    10,
+            MetricRequestDatetime: time.Now(),
+            EventFieldString:      "event extra string value",
+            EventFieldQuantity:    100,
+            EventFieldBool:        false,
         },
     )
     """
@@ -161,6 +189,58 @@ def test_run_logging(tmp_path):
     assert "glean-server-event" == logged_output["Type"]
     assert "glean.test" == fields["document_namespace"]
     assert "events" == fields["document_type"]
+    assert "1" == fields["document_version"]
+    assert "glean-test/1.0" == fields["user_agent"]
+
+    schema_url = (
+        "https://raw.githubusercontent.com/mozilla-services/"
+        "mozilla-pipeline-schemas/main/"
+        "schemas/glean/glean/glean.1.schema.json"
+    )
+
+    input = io.StringIO(payload)
+    output = io.StringIO()
+    assert (
+        validate_ping.validate_ping(input, output, schema_url=schema_url) == 0
+    ), output.getvalue()
+
+
+@pytest.mark.go_dependency
+def test_run_logging_custom_ping(tmp_path):
+    glean_module_path = tmp_path / "glean"
+
+    translate.translate(
+        [
+            ROOT / "data" / "go_server_custom_ping_only_metrics.yaml",
+            ROOT / "data" / "go_server_custom_ping_only_pings.yaml"
+        ],
+        "go_server",
+        glean_module_path,
+    )
+
+    code = """
+    logger.RecordPingServerTelemetryScenarioOne(
+        glean.RequestInfo{
+            UserAgent: "glean-test/1.0",
+            IpAddress: "127.0.0.1",
+        },
+        glean.PingServerTelemetryScenarioOne{
+            MetricName:            "string value",
+            MetricRequestBool:     true,
+            MetricRequestCount:    20,
+            MetricRequestDatetime: time.Now(),
+        },
+    )
+    """
+
+    logged_output = run_logger(tmp_path, code)
+    logged_output = json.loads(logged_output)
+    fields = logged_output["Fields"]
+    payload = fields["payload"]
+
+    assert "glean-server-event" == logged_output["Type"]
+    assert "glean.test" == fields["document_namespace"]
+    assert "server-telemetry-scenario-one" == fields["document_type"]
     assert "1" == fields["document_version"]
     assert "glean-test/1.0" == fields["user_agent"]