+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add Go log outputter support for custom pings (`go_server`) ([???](https://github.com/mozilla/glean_parser/pull/???))
+
 ## 15.0.1
 
 - Rust codegen: use correctly named parameter for events without extras ([#750](https://github.com/mozilla/glean_parser/pull/750))